+++ CHANGELOG.md
@@ -29,6 +29,7 @@ This release does not contain security updates.
 - `OSLC4Net.Core` requires .NET 10 to be able to use the `[Experimental]` annotation.
 - `OSLC4Net.Client` requires .NET 10.
 - `OslcClient` now has a `DefaultRequestParams` property for default request parameters configuration.
+- ❗️ `SignedByteNode` (which corresponds to `xsd:byte`) is now parsed as C# `sbyte` (signed byte) instead of `byte`.
 
 ### Deprecated